Small fixes, implemented js parser using 'MyParser'
