+++ src/main/scala/chapter_9/ParseError.scala
@@ -1,17 +1,17 @@
 package chapter_9
 
-case class ParseError(stack: List[(Location, String)]) {
+case class ParseError(stack: List[(Location, String)] = Nil) {
 
   def push(loc: Location, msg: String): ParseError =
     copy(stack = (loc, msg) :: stack)
 
   def label[A](s: String): ParseError =
-    ParseError(latestLoc.map((_,s)).toList)
+    ParseError(latestLoc.map((_, s)).toList)
 
   def latestLoc: Option[Location] =
     latest map (_._1)
 
-  def latest: Option[(Location,String)] =
+  def latest: Option[(Location, String)] =
     stack.lastOption
 
 }
